GraphiteBack/requirements.txt
Artists==0.0.2
asgiref==3.4.1
blank==0.1.0
cached-property==1.5.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
comments==1.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
decorator==5.1.0
defusedxml==0.7.1
Deprecated==1.2.13
Django==3.2.9
django-cors-headers==3.10.0
django-extensions==3.1.5
django-filter==21.1
django-grappelli==2.15.3
django-oauth-toolkit==1.5.0
django-templated-mail==1.1.1
django-url-filter==0.3.15
djangorestframework==3.12.4
djangorestframework-filters==1.0.0.dev2
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-extensions==0.7.1
enum-compat==0.0.3
envparse==0.2.0
gunicorn==20.1.0
h11==0.12.0
httptools==0.2.0
idna==3.3
itypes==1.2.0
Jinja2==3.0.3
jwcrypto==1.0
MarkupSafe==2.0.1
nickname==0.0.0.7
null==0.6.1
oauth2-provider==0.0
oauthlib==3.1.1
owner==0.0.10
Pillow==8.4.0
psycopg2==2.9.1
pycparser==2.21
pydot==1.4.2
PyJWT==2.3.0
pyparsing==3.0.6
python-dotenv==0.19.1
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.0
self==2020.12.3
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.7
uvicorn==0.15.0
watchgod==0.7
websockets==10.1
whitenoise==5.3.0
wrapt==1.13.3
